Added usage of the `menu-maker` utility.
This change makes it so that we can create individual
markdown files for each pre-sales offering and generate
a "menu" from all of those files.

The `build.sh` script has been added to reliably run
this script. Next step is to add a pre-commit hook
that runs this script so that the menu is always up
to date.